+++ packages/swapper/src/swappers/JupiterSwapper/utils/helpers.ts
@@ -1,16 +1,20 @@
+import { BorshInstructionCoder } from '@coral-xyz/anchor'
 import { ASSOCIATED_PROGRAM_ID, TOKEN_PROGRAM_ID } from '@coral-xyz/anchor/dist/cjs/utils/token'
-import type { QuoteResponse, SwapInstructionsResponse } from '@jup-ag/api'
+import type { Instruction, QuoteResponse, SwapInstructionsResponse } from '@jup-ag/api'
 import type { ChainId } from '@shapeshiftoss/caip'
-import { fromAssetId } from '@shapeshiftoss/caip'
-import type { KnownChainIds } from '@shapeshiftoss/types'
+import { fromAssetId, solAssetId, wrappedSolAssetId } from '@shapeshiftoss/caip'
+import type { ChainAdapter } from '@shapeshiftoss/chain-adapters/src/solana'
+import type { Asset, KnownChainIds } from '@shapeshiftoss/types'
 import { bnOrZero } from '@shapeshiftoss/utils'
 import type { Result } from '@sniptt/monads'
 import type { Connection } from '@solana/web3.js'
 import { PublicKey, SystemProgram, TransactionInstruction } from '@solana/web3.js'
-import type { AxiosResponse } from 'axios'
+import type { AxiosError, AxiosResponse } from 'axios'
 
 import type { SwapErrorRight } from '../../../types'
+import { referralIdl } from '../idls/referral'
 import {
+  JUPITER_AFFILIATE_CONTRACT_ADDRESS,
   JUPITER_REFERALL_FEE_PROJECT_ACCOUNT,
   jupiterSupportedChainIds,
   PDA_ACCOUNT_CREATION_COST,
@@ -40,6 +44,17 @@ type GetJupiterSwapArgs = {
   toAddress?: string
   useSharedAccounts: boolean
   feeAccount: string | undefined
+}
+
+type CreateInstructionsParams = {
+  priceResponse: QuoteResponse
+  sendAddress: string
+  receiveAddress?: string
+  affiliateBps: string
+  buyAsset: Asset
+  sellAsset: any
+  adapter: ChainAdapter
+  jupiterUrl: string
 }
 
 export const getJupiterPrice = ({
@@ -176,4 +191,123 @@ export const calculateAccountCreationCosts = (instructions: TransactionInstructi
   }
 
   return totalCost.toString()
+}
+
+export const createSwapInstructions = async ({
+  priceResponse,
+  sendAddress,
+  receiveAddress,
+  affiliateBps,
+  buyAsset,
+  sellAsset,
+  adapter,
+  jupiterUrl,
+}: CreateInstructionsParams): Promise<{
+  instructions: TransactionInstruction[]
+  addressLookupTableAddresses: string[]
+}> => {
+  const isCrossAccountTrade = receiveAddress ? receiveAddress !== sendAddress : false
+
+  const buyAssetAddress =
+    buyAsset.assetId === solAssetId
+      ? fromAssetId(wrappedSolAssetId).assetReference
+      : fromAssetId(buyAsset.assetId).assetReference
+
+  const sellAssetAddress =
+    sellAsset.assetId === solAssetId
+      ? fromAssetId(wrappedSolAssetId).assetReference
+      : fromAssetId(sellAsset.assetId).assetReference
+
+  const contractAddress =
+    buyAsset.assetId === solAssetId ? undefined : fromAssetId(buyAsset.assetId).assetReference
+
+  const { instruction: createTokenAccountInstruction, destinationTokenAccount } =
+    contractAddress && isCrossAccountTrade
+      ? await adapter.createAssociatedTokenAccountInstruction({
+          from: sendAddress,
+          to: receiveAddress!,
+          tokenId: contractAddress,
+        })
+      : { instruction: undefined, destinationTokenAccount: undefined }
+
+  const [buyAssetReferralPubKey] = PublicKey.findProgramAddressSync(
+    [
+      Buffer.from('referral_ata'),
+      new PublicKey(SHAPESHIFT_JUPITER_REFERRAL_KEY).toBuffer(),
+      new PublicKey(buyAssetAddress).toBuffer(),
+    ],
+    new PublicKey(JUPITER_AFFILIATE_CONTRACT_ADDRESS),
+  )
+
+  const [sellAssetReferralPubKey] = PublicKey.findProgramAddressSync(
+    [
+      Buffer.from('referral_ata'),
+      new PublicKey(SHAPESHIFT_JUPITER_REFERRAL_KEY).toBuffer(),
+      new PublicKey(sellAssetAddress).toBuffer(),
+    ],
+    new PublicKey(JUPITER_AFFILIATE_CONTRACT_ADDRESS),
+  )
+
+  const instructionData = new BorshInstructionCoder(referralIdl).encode(
+    'initializeReferralTokenAccount',
+    {},
+  )
+
+  const { instruction: feeAccountInstruction, tokenAccount } =
+    await getFeeTokenAccountAndInstruction({
+      feePayerPubKey: new PublicKey(sendAddress),
+      buyAssetReferralPubKey,
+      sellAssetReferralPubKey,
+      programId: new PublicKey(JUPITER_AFFILIATE_CONTRACT_ADDRESS),
+      instructionData,
+      buyTokenId: buyAssetAddress,
+      sellTokenId: sellAssetAddress,
+      connection: adapter.getConnection(),
+    })
+
+  const maybeSwapResponse = await getJupiterSwapInstructions({
+    apiUrl: jupiterUrl,
+    fromAddress: sendAddress,
+    toAddress: isCrossAccountTrade ? destinationTokenAccount?.toString() : undefined,
+    rawQuote: priceResponse,
+    useSharedAccounts: priceResponse.routePlan.length > 1 && isCrossAccountTrade ? true : false,
+    feeAccount: affiliateBps !== '0' ? tokenAccount?.toString() : undefined,
+  })
+
+  if (maybeSwapResponse.isErr()) {
+    const error = maybeSwapResponse.unwrapErr()
+    const cause = error.cause as AxiosError<any, any>
+    throw Error(cause.response!.data.detail)
+  }
+
+  const { data: swapResponse } = maybeSwapResponse.unwrap()
+
+  const convertJupiterInstruction = (instruction: Instruction): TransactionInstruction => ({
+    ...instruction,
+    keys: instruction.accounts.map(account => ({
+      ...account,
+      pubkey: new PublicKey(account.pubkey),
+    })),
+    data: Buffer.from(instruction.data, 'base64'),
+    programId: new PublicKey(instruction.programId),
+  })
+
+  const instructions: TransactionInstruction[] = [
+    ...swapResponse.setupInstructions.map(convertJupiterInstruction),
+    convertJupiterInstruction(swapResponse.swapInstruction),
+  ]
+
+  if (feeAccountInstruction && affiliateBps !== '0') {
+    instructions.unshift(feeAccountInstruction)
+  }
+
+  if (createTokenAccountInstruction) {
+    instructions.unshift(createTokenAccountInstruction)
+  }
+
+  if (swapResponse.cleanupInstruction) {
+    instructions.push(convertJupiterInstruction(swapResponse.cleanupInstruction))
+  }
+
+  return { instructions, addressLookupTableAddresses: swapResponse.addressLookupTableAddresses }
 }
